Refactor project structure based on Clean Architecure
Modified `launchSettings.json` for new profiles.Updated `back.sln` to include new projects and remove `back`.
