+++ back/Properties/launchSettings.json
@@ -6,7 +6,9 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       },
       "dotnetRunMessages": true,
-      "applicationUrl": "http://localhost:5115"
+      "applicationUrl": "http://localhost:5115",
+      "launchBrowser": true,
+      "launchUrl": "scalar/v1"
     },
     "https": {
       "commandName": "Project",
@@ -14,7 +16,9 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       },
       "dotnetRunMessages": true,
-      "applicationUrl": "https://localhost:7239;http://localhost:5115"
+      "applicationUrl": "https://localhost:7239;http://localhost:5115",
+      "launchBrowser": true,
+      "launchUrl": "scalar/v1"
     },
     "Container (Dockerfile)": {
       "commandName": "Docker",
@@ -25,7 +29,23 @@
       },
       "publishAllPorts": true,
       "useSSL": true
+    },
+    "IIS Express": {
+      "commandName": "IISExpress",
+      "launchBrowser": true,
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      },
+      "launchUrl": "scalar/v1"
     }
   },
-  "$schema": "https://json.schemastore.org/launchsettings.json"
+  "$schema": "https://json.schemastore.org/launchsettings.json",
+  "iisSettings": {
+    "windowsAuthentication": false,
+    "anonymousAuthentication": true,
+    "iisExpress": {
+      "applicationUrl": "http://localhost:51016/",
+      "sslPort": 44329
+    }
+  }
 }
