+++ back/PublicApi/Dockerfile
@@ -12,19 +12,19 @@ EXPOSE 8081
 FROM mcr.microsoft.com/dotnet/sdk:9.0 AS build
 ARG BUILD_CONFIGURATION=Release
 WORKDIR /src
-COPY ["back.csproj", "."]
-RUN dotnet restore "./back.csproj"
+COPY ["PublicApi/PublicApi.csproj", "PublicApi/"]
+RUN dotnet restore "./PublicApi/PublicApi.csproj"
 COPY . .
-WORKDIR "/src/."
-RUN dotnet build "./back.csproj" -c $BUILD_CONFIGURATION -o /app/build
+WORKDIR "/src/PublicApi"
+RUN dotnet build "./PublicApi.csproj" -c $BUILD_CONFIGURATION -o /app/build
 
 # This stage is used to publish the service project to be copied to the final stage
 FROM build AS publish
 ARG BUILD_CONFIGURATION=Release
-RUN dotnet publish "./back.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
+RUN dotnet publish "./PublicApi.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
 
 # This stage is used in production or when running from VS in regular mode (Default when not using the Debug configuration)
 FROM base AS final
 WORKDIR /app
 COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "back.dll"]
+ENTRYPOINT ["dotnet", "PublicApi.dll"]
