+++ back/PublicApi/PublicApi.http
@@ -1,0 +1,6 @@
+@PublicApi_HostAddress = http://localhost:5095
+
+GET {{PublicApi_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
