--- back/back.http
@@ -1,6 +1,0 @@
-@back_HostAddress = http://localhost:5115
-
-GET {{back_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
