+++ back/back.sln
@@ -1,9 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.12.35527.113 d17.12
+VisualStudioVersion = 17.12.35527.113
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "back", "back.csproj", "{240ED8B3-24BE-4EC3-9E23-AAA094F75C09}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PublicApi", "PublicApi\PublicApi.csproj", "{0FB57AC9-1F02-4340-A718-975AE629D01F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AppCore", "AppCore\AppCore.csproj", "{44FCED54-8CDD-456C-AD42-81BD819EF6F6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure", "Infrastructure\Infrastructure.csproj", "{D465DFD5-99B0-4984-B7B9-E4CE4F7ECF76}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +15,18 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{240ED8B3-24BE-4EC3-9E23-AAA094F75C09}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{240ED8B3-24BE-4EC3-9E23-AAA094F75C09}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{240ED8B3-24BE-4EC3-9E23-AAA094F75C09}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{240ED8B3-24BE-4EC3-9E23-AAA094F75C09}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0FB57AC9-1F02-4340-A718-975AE629D01F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0FB57AC9-1F02-4340-A718-975AE629D01F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0FB57AC9-1F02-4340-A718-975AE629D01F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0FB57AC9-1F02-4340-A718-975AE629D01F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{44FCED54-8CDD-456C-AD42-81BD819EF6F6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{44FCED54-8CDD-456C-AD42-81BD819EF6F6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{44FCED54-8CDD-456C-AD42-81BD819EF6F6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{44FCED54-8CDD-456C-AD42-81BD819EF6F6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D465DFD5-99B0-4984-B7B9-E4CE4F7ECF76}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D465DFD5-99B0-4984-B7B9-E4CE4F7ECF76}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D465DFD5-99B0-4984-B7B9-E4CE4F7ECF76}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D465DFD5-99B0-4984-B7B9-E4CE4F7ECF76}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
